+++ src/main/java/ru/job4j/cinema/repository/GenreRepository.java
@@ -6,5 +6,5 @@ import java.util.Optional;
 
 public interface GenreRepository {
 
-    Optional<Genre> findById(int id);
+    Optional<Genre> findById(Long id);
 }
