+++ src/main/java/ru/job4j/cinema/repository/Sql2oFilmRepository.java
@@ -19,7 +19,7 @@ public class Sql2oFilmRepository implements FilmRepository {
     }
 
     @Override
-    public Optional<Film> findById(int id) {
+    public Optional<Film> findById(Long id) {
         try (var connection = sql2o.open()) {
             var query = connection.createQuery("SELECT * FROM films WHERE id = :id")
                     .addParameter("id", id);
